+++ client/web/src/enterprise/insights/core/hooks/live-preview-insight/types.ts
@@ -1,0 +1,12 @@
+export enum LivePreviewStatus {
+    Intact,
+    Loading,
+    Error,
+    Data,
+}
+
+export type State<D> =
+    | { status: LivePreviewStatus.Data; data: D }
+    | { status: LivePreviewStatus.Error; error: Error }
+    | { status: LivePreviewStatus.Loading }
+    | { status: LivePreviewStatus.Intact }