+++ enterprise/cmd/worker/internal/batches/dbstore.go
@@ -7,6 +7,7 @@ import (
 
 	workerdb "github.com/sourcegraph/sourcegraph/cmd/worker/shared/init/db"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/batches/store"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
 	"github.com/sourcegraph/sourcegraph/internal/encryption/keyring"
 	"github.com/sourcegraph/sourcegraph/internal/memo"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
@@ -36,11 +37,11 @@ var initStore = memo.NewMemoizedConstructor(func() (*store.Store, error) {
 })
 
 // InitReconcilerWorkerStore initializes and returns a dbworker.Store instance for the reconciler worker.
-func InitReconcilerWorkerStore() (dbworkerstore.Store, error) {
+func InitReconcilerWorkerStore() (dbworkerstore.Store[*types.Changeset], error) {
 	return initReconcilerWorkerStore.Init()
 }
 
-var initReconcilerWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store, error) {
+var initReconcilerWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store[*types.Changeset], error) {
 	observationContext := &observation.Context{
 		Logger:     log.Scoped("store.reconciler", "reconciler worker store"),
 		Tracer:     &trace.Tracer{TracerProvider: otel.GetTracerProvider()},
@@ -56,11 +57,11 @@ var initReconcilerWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstor
 })
 
 // InitBulkOperationWorkerStore initializes and returns a dbworker.Store instance for the bulk operation processor worker.
-func InitBulkOperationWorkerStore() (dbworkerstore.Store, error) {
+func InitBulkOperationWorkerStore() (dbworkerstore.Store[*types.ChangesetJob], error) {
 	return initBulkOperationWorkerStore.Init()
 }
 
-var initBulkOperationWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store, error) {
+var initBulkOperationWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store[*types.ChangesetJob], error) {
 	observationContext := &observation.Context{
 		Logger:     log.Scoped("store.bulk_ops", "bulk operation worker store"),
 		Tracer:     &trace.Tracer{TracerProvider: otel.GetTracerProvider()},
@@ -76,11 +77,11 @@ var initBulkOperationWorkerStore = memo.NewMemoizedConstructor(func() (dbworkers
 })
 
 // InitBatchSpecWorkspaceExecutionWorkerStore initializes and returns a dbworkerstore.Store instance for the batch spec workspace execution worker.
-func InitBatchSpecWorkspaceExecutionWorkerStore() (dbworkerstore.Store, error) {
+func InitBatchSpecWorkspaceExecutionWorkerStore() (dbworkerstore.Store[*types.BatchSpecWorkspaceExecutionJob], error) {
 	return initBatchSpecWorkspaceExecutionWorkerStore.Init()
 }
 
-var initBatchSpecWorkspaceExecutionWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store, error) {
+var initBatchSpecWorkspaceExecutionWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store[*types.BatchSpecWorkspaceExecutionJob], error) {
 	observationContext := &observation.Context{
 		Logger:     log.Scoped("store.execution", "the batch spec workspace execution worker store"),
 		Tracer:     &trace.Tracer{TracerProvider: otel.GetTracerProvider()},
@@ -96,11 +97,11 @@ var initBatchSpecWorkspaceExecutionWorkerStore = memo.NewMemoizedConstructor(fun
 })
 
 // InitBatchSpecResolutionWorkerStore initializes and returns a dbworker.Store instance for the batch spec workspace resolution worker.
-func InitBatchSpecResolutionWorkerStore() (dbworkerstore.Store, error) {
+func InitBatchSpecResolutionWorkerStore() (dbworkerstore.Store[*types.BatchSpecResolutionJob], error) {
 	return initBatchSpecResolutionWorkerStore.Init()
 }
 
-var initBatchSpecResolutionWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store, error) {
+var initBatchSpecResolutionWorkerStore = memo.NewMemoizedConstructor(func() (dbworkerstore.Store[*types.BatchSpecResolutionJob], error) {
 	observationContext := &observation.Context{
 		Logger:     log.Scoped("store.batch_spec_resolution", "the batch spec resolution worker store"),
 		Tracer:     &trace.Tracer{TracerProvider: otel.GetTracerProvider()},