+++ internal/database/webhooks_test.go
@@ -22,8 +22,10 @@ import (
 )
 
 const (
-	testSecret = "my secret"
-	testURN    = "https://github.com"
+	testSecret        = "my secret"
+	testURN           = "https://github.com"
+	githubWebhookName = "GitHub webhook"
+	gitlabWebhookName = "GitLab webhook"
 )
 
 func TestWebhookCreate(t *testing.T) {
@@ -42,12 +44,13 @@ func TestWebhookCreate(t *testing.T) {
 			codeHostURL := "https://github.com/"
 			encryptedSecret := types.NewUnencryptedSecret(testSecret)
 
-			created, err := store.Create(ctx, kind, codeHostURL, 0, encryptedSecret)
+			created, err := store.Create(ctx, githubWebhookName, kind, codeHostURL, 0, encryptedSecret)
 			assert.NoError(t, err)
 
 			// Check that the calculated fields were correctly calculated.
 			assert.NotZero(t, created.ID)
 			assert.NotZero(t, created.UUID)
+			assert.Equal(t, githubWebhookName, created.Name)
 			assert.Equal(t, kind, created.CodeHostKind)
 			assert.Equal(t, codeHostURL, created.CodeHostURN.String())
 			assert.Equal(t, int32(0), created.CreatedByUserID)
@@ -80,13 +83,14 @@ func TestWebhookCreate(t *testing.T) {
 		kind := extsvc.KindGitHub
 		codeHostURL := "https://github.com/"
 
-		created, err := store.Create(ctx, kind, codeHostURL, 0, nil)
+		created, err := store.Create(ctx, githubWebhookName, kind, codeHostURL, 0, nil)
 		assert.NoError(t, err)
 
 		// Check that the calculated fields were correctly calculated.
 		assert.NotZero(t, created.ID)
 		assert.NotZero(t, created.UUID)
 		assert.NoError(t, err)
+		assert.Equal(t, githubWebhookName, created.Name)
 		assert.Equal(t, kind, created.CodeHostKind)
 		assert.Equal(t, codeHostURL, created.CodeHostURN.String())
 		assert.Equal(t, int32(0), created.CreatedByUserID)
@@ -131,7 +135,7 @@ func TestWebhookCreate(t *testing.T) {
 	t.Run("with bad key", func(t *testing.T) {
 		store := db.Webhooks(&et.BadKey{Err: errors.New("some error occurred, sorry")})
 
-		_, err := store.Create(ctx, extsvc.KindGitHub, "https://github.com", 0, types.NewUnencryptedSecret("very secret (not)"))
+		_, err := store.Create(ctx, "name", extsvc.KindGitHub, "https://github.com", 0, types.NewUnencryptedSecret("very secret (not)"))
 		assert.Error(t, err)
 		assert.Equal(t, "encrypting secret: some error occurred, sorry", err.Error())
 	})
@@ -281,7 +285,7 @@ func createWebhookWithActorUID(ctx context.Context, t *testing.T, actorUID int32
 	kind := extsvc.KindGitHub
 	encryptedSecret := types.NewUnencryptedSecret(testSecret)
 
-	created, err := store.Create(ctx, kind, testURN, actorUID, encryptedSecret)
+	created, err := store.Create(ctx, githubWebhookName, kind, testURN, actorUID, encryptedSecret)
 	assert.NoError(t, err)
 	return created
 }
@@ -333,8 +337,10 @@ func TestWebhookList(t *testing.T) {
 		assert.Len(t, webhooks, 2)
 		assert.Equal(t, webhooks[0].ID, int32(2))
 		assert.Equal(t, webhooks[0].CodeHostKind, extsvc.KindGitHub)
+		assert.Equal(t, webhooks[0].Name, githubWebhookName)
 		assert.Equal(t, webhooks[1].ID, int32(3))
 		assert.Equal(t, webhooks[1].CodeHostKind, extsvc.KindGitLab)
+		assert.Equal(t, webhooks[1].Name, gitlabWebhookName)
 	})
 
 	t.Run("with pagination and filtering by code host kind", func(t *testing.T) {
@@ -388,9 +394,9 @@ func createTestWebhooks(ctx context.Context, t *testing.T, store WebhookStore) (
 		var err error
 		if i%3 == 0 {
 			numGitlabHooks++
-			_, err = store.Create(ctx, extsvc.KindGitLab, fmt.Sprintf("http://instance-%d.github.com", i), 0, encryptedSecret)
+			_, err = store.Create(ctx, gitlabWebhookName, extsvc.KindGitLab, fmt.Sprintf("http://instance-%d.github.com", i), 0, encryptedSecret)
 		} else {
-			_, err = store.Create(ctx, extsvc.KindGitHub, fmt.Sprintf("http://instance-%d.gitlab.com", i), 0, encryptedSecret)
+			_, err = store.Create(ctx, githubWebhookName, extsvc.KindGitHub, fmt.Sprintf("http://instance-%d.gitlab.com", i), 0, encryptedSecret)
 		}
 		assert.NoError(t, err)
 	}
@@ -416,7 +422,7 @@ func TestGetByID(t *testing.T) {
 	assert.Nil(t, webhook)
 
 	// Test that existent webhook cannot be found
-	createdWebhook, err := store.Create(ctx, extsvc.KindGitHub, "https://github.com", 0, types.NewUnencryptedSecret("very secret (not)"))
+	createdWebhook, err := store.Create(ctx, githubWebhookName, extsvc.KindGitHub, "https://github.com", 0, types.NewUnencryptedSecret("very secret (not)"))
 	assert.NoError(t, err)
 
 	webhook, err = store.GetByID(ctx, createdWebhook.ID)
@@ -425,6 +431,7 @@ func TestGetByID(t *testing.T) {
 	assert.Equal(t, webhook.ID, createdWebhook.ID)
 	assert.Equal(t, webhook.UUID, createdWebhook.UUID)
 	assert.Equal(t, webhook.Secret, createdWebhook.Secret)
+	assert.Equal(t, webhook.Name, createdWebhook.Name)
 	assert.Equal(t, webhook.CodeHostKind, createdWebhook.CodeHostKind)
 	assert.Equal(t, webhook.CodeHostURN.String(), createdWebhook.CodeHostURN.String())
 	assert.Equal(t, webhook.CreatedAt, createdWebhook.CreatedAt)
@@ -446,7 +453,7 @@ func TestGetByUUID(t *testing.T) {
 	assert.Nil(t, webhook)
 
 	// Test that existent webhook cannot be found
-	createdWebhook, err := store.Create(ctx, extsvc.KindGitHub, "https://github.com", 0, types.NewUnencryptedSecret("very secret (not)"))
+	createdWebhook, err := store.Create(ctx, githubWebhookName, extsvc.KindGitHub, "https://github.com", 0, types.NewUnencryptedSecret("very secret (not)"))
 	assert.NoError(t, err)
 
 	webhook, err = store.GetByUUID(ctx, createdWebhook.UUID)
@@ -455,6 +462,7 @@ func TestGetByUUID(t *testing.T) {
 	assert.Equal(t, webhook.ID, createdWebhook.ID)
 	assert.Equal(t, webhook.UUID, createdWebhook.UUID)
 	assert.Equal(t, webhook.Secret, createdWebhook.Secret)
+	assert.Equal(t, webhook.Name, createdWebhook.Name)
 	assert.Equal(t, webhook.CodeHostKind, createdWebhook.CodeHostKind)
 	assert.Equal(t, webhook.CodeHostURN.String(), createdWebhook.CodeHostURN.String())
 	assert.Equal(t, webhook.CreatedAt, createdWebhook.CreatedAt)